+++ SETOCryptomatorCryptor/SETOCryptorV5.h
@@ -1,0 +1,13 @@
+//
+//  SETOCryptorV5.h
+//  SETOCryptomatorCryptor
+//
+//  Created by Tobias Hagemann on 02/09/16.
+//  Copyright © 2015-2016 setoLabs. All rights reserved.
+//
+
+#import "SETOCryptor.h"
+
+@interface SETOCryptorV5 : SETOCryptor
+
+@end