+++ RELEASE NOTES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## Version 1.2.0
+
+We've decided to disable file size obfuscation. File sizes can be determined in O(1) instead of having to read and decrypt the file header. This allows showing file sizes in the directory listing without having to download each file first.
+
+- Added vault version 5 compatibility.
+- Added `+[SETOCryptorProvider cleartextSizeFromCiphertextSize:withCryptor:]` and `+[SETOCryptorProvider ciphertextSizeFromCleartextSize:withCryptor:]` methods for version 5 cryptors.
+
 ## Version 1.1.4
 
 - `SETOCryptor` version is now an accessible property.